Updates custom functions to use proper importing of modules within the new 'my_mods' package
